moved spam config to config.go
